MB-25342: Add a callback to discover if a key is added/replaced
Add a new couchstore_save_docs method which allows the caller to
specify a callback, the callback is invoked by the btree "builder"
with the key, its seqno, a user context and an enum value indicating
if the key was added vs replaced.

For collections this functionality is intended to assist when we don't
have all keys in cached (i.e. full eviction) and wish to keep count
of collection cardinality. I.e. when couchstore tells us it added a
key, we shall increment the key's collection statistics.

Change-Id: I71e3d99e2d5cb01e57f480ba51d63f25669a42cb